slack-bolt-app/pyproject.toml
[tool.poetry]
name = "slack-bolt-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
slack-bolt = "^1.18.0"
python-dotenv = "^1.0.0"
httpx = "^0.25.0"
pydash = "^7.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "src.app:run"